Deleted PredicateCollector
The implementation I initially had in mind for this class did not fit the model I actually am looking for. Developing further into this direction would have lead to alot of wasted time, I think. So for now I deleted this class. It will come back in another form and place eventually, as I still need to access the predicates of a machine for training data generation.