Speed up in finding problematic machines.
As in previous commits mentioned there are problems with some machines, causing ProB2 to simply crash everything with an uncatchable exception. To find those files I now finished a system, that (a) scans the desired directory for subdirectories, (b) lists those subdirs line-wise in a file, and (c) reads this file line by line, generating feature data from the machines inside the given directory and deletes the respective line from the file as soon as the directory has been searched for feature data correctly.

This allows to let the training set generation run through the directories, and on a crash simply resume in the last directory it was looking, rather than starting from the start as before. This saves alot of time, as it takes roughly 20 minutes to go through all directories until the first problem causing machine is encountered, and I am not even half way through.