Added new features.
New features revolve around the size of domains of variables. There are some problems with it right now, as the BParser called for infering this information fails to comprehend NATURAL or INTEGER as predefined sets. Instead they are understood as identifiers on their own.

It remains to find a way to get the information from the parser as needed.