Added option to use gradle run with command line arguments.
Gradle does not natively support forwarding of command line arguments to the application when calling gradle run. Instead it has to be set manually, and looks very ugly, too.

To call the NeuroBCli with commandline argments, call the following: gradle run -PappArgs="['arg1', 'arg2', ...]"
Note, that there are quotation marks around the brackets "[...]" and around the single arguments as well ('arg1', 'arg2'). This is rather ugly, but at least it works.