Changed labeling behaviour and added timeout.
Previously the labeling was 1 for solvable predicates, 0 for undecidable ones and -1 for contradictable ones. As we are not interested in whether a predicate has a solution or a contradiction, but if the solver computes a result or not, the new labeling is 1 for result found and 0 for no result found.