Version 0.10.0
Logical change in the labeling of training data. Now we only decide if a given solver can find a solution or contradiction, or fails to decide the predicate. This is, as NeuroB (for now) shall pick a solver to determine a solution, and thus has no need to distinguish between solvable and contradictable constraints (but rather has them both as a single class: decidable by given solver).