The neural net is now trainable via the NeuroB class.
There were some problems with the multilabeling approach, as the Record Reader of Deeplearning4j only takes a single lable index in the constructor I used.
This is now bypassed by using a different constructor, which is limited to the use of regression as it seems. For now the model can be trained, but the results are as bad as they could be: The net seems to only label to a single value.