Added utility to generate CSV with only interesting entries.
Some training vectors map to a 3-way classification with labeling vector 1,1,1 or 0,0,0. As for now this is the majority of data, but does not help with a decision problem, the csv generation can now be called with the --ignoreEquallyLabeledEntries argument, that does not take that data into the csv file.

Note: This strongly encurrages the neural net to map the predicates found onto different labelings, but leaves it unfamiliar with situations in which indeed the data should be mapped onto all classes.