Starting to get rid of DefaultTrainingDataCollector.
The class should no longer have a real purpose, but instead be replaced by classed with more meaningfull names one can enjoyfully work with. It shall be replaced by SolverClassificationDataCollector, which does exactly the same for now.

Also some of the code got outsourced to a new helper class, FormulaGenerator, that handles the construction and rearrangements of formulas to be used by different data collectors that come in the future.