Force exiting from Cli after the command is done.
For now the problem remains, that some machines do not load properly into Prob2, resulting in a started ProBCli process I have no handle to so I cannot close it from whithin the code. This is a problem of ProB2.

As now multiple net are present, each needs a trainingset and I want to create them atleast sequentially, I end up with alot of ProBCli processes that do not terminate. System.exit at the very end of my code should ensure, that the Cli exits none the less and takes those processes with it.