Made an java interface for features used by NeuroB.
Part of restructuring the API for neural nets.

As the idea is, to have a more plug-and-play like structure regarding features to be generated and labels to be matched against them, those two parts should become more cleanly separated. The features interface provides a small API with the most basic methods all feature generators should provide.