Generating train and test sets from data.
Instead of one data.csv now the training set generator creates two csv files: train_data.csv and test_data.csv. The training file can be used for the training of the network, whilst the test data should only be used to evaluate a already trained network.

Also, the NeuroBNet class now handles its own normaliser. This is maybe still not implemented perfectly, as the usage seems to be a bit clumsy.