Changed basic model structure of NeuroBNet.
The net now uses tanh as activation function for hidden layers and negative log-likelyhood as loss function in the output layer.
Also I activated normalisation/preprocessing of data again, so the tanh does not result in vanishing gradients.

This should counter a problem with the score in training being 0 all the time, but still things seem to be off, as the score does not decrease e.g. for the ProB classification.

Other issues are that the evaluation after training seems to be buggy, as always the same results are shown: accuracy of 0.027~; precision, recall, and f1score of 0.5. This still needs to be looked into at a future point in time.