Fix in solver selection regarding cases in which not all solvers failed.
For the predicate x<y&y<x, ProB and KodKod fail (timer value: -1), but Z3 succeeds. As not all failed (-1), not 0 is returned as class, but 1 (ProB).

This bug was due to taking the argMin to get the fastest solver, ignoring the possibility that one of them could still have failed. This is now corrected.