Catch block for generall Exception class in training set generation and in predicate dumper.
Solves mysterious problem of randomly stopping threads.

It happened, that each thread would eventually find a file, where the second generated formula would cause a EvaliationException. As this is an unchecked exception, this resulted in the thread stopping, but leaving the StateSpace open. Bad. Result was all theads being open but doing nothing, and as the excpetion could not be logged (I did not now it could be thrown) it confused the hell out of me.
Solution was to simply put an additional catch(Exception e)-block at the end of formulae evaluation. Now each thread should run through properly and close opened state spaces.