Changed some created formulae.
First, P=>Q was not a predicate, that yielded much insight, as the solvers simply tried to falsify P. Changed it to ~P=>Q, so solvers have to proof P.

Second, I took out all the P<=>Q formulae, that kind of had similar issues. Predicates where this would be interesting probably are edge cases anyway.