Generalised TrainingSetGenerator.
Now the trainingset generator takes a TrainingDataGenerator object to create the training files. This is partially hidden in the training package, so that an enduser does not have to know about it.

The NeuroBNet method to get a trainingset generator simply makes use of a change for the LabelGenerator interface, which now provides a method to couple a FeatureGenerator with it to provide a TrainingDataGenerator.

The name is not really distinct from TrainingSetGenerator, but it should be not much of a problem, as the user should actually just use the NeuroB class, which even hides the TrainingSetGenerator.