Getter for training data generator was refactored to be part of feature generator instead of label generator.
As the feature generator may hold utility to additionally create image features or sequential features (NYI), it is on this interface to decide what training data generator to use, not on the label generator.

Added PredicateTrainingImageGenerator for convolutional image predicates to use. This class does nothing different than PredicateTrainingDataGenerator for now, but extends it and shall in future be responsible to generate hierarchies of images instead of nbtrain files.