Convolution features are now images that carry the label in the name rather than their parent directory.
Previously the labelling of an image was determined by the directory it was placed in. This only seemed useful for classification problems. Tackling regression with convolution, on the other hand, would have lead to an uncontrollable amount of directories needed. As images now have the label as part of their name, no excessive amount of directories has to be created and regression is no problem anymore.