Adapt NeuroBModelSpace methods to a bug encountered in DL4J Arbiter
Main problem is the following:
Internally the goal is to set up multiple layers with e.g. varying amounts of neurons.
The amount of neurons per layer is limited by the same pair of bounds (lower, upper).
This pair of bounds spans the parameter space. Now, Arbiter internally rationalises those parameter
spaces away for any hidden layer but the first one, as all parameter spaces equal one another
(same lower and upper bound) though they are not the same objects.

The workaround for now is to simply use the same parameter space for each layer configuration,
with the result of having all parameters of the same size.