Remove IntelliJ specific files
Initially I checked those in, as it was suggested by the Jetbrains Doc.
However, it appeares to be mildy annoying having to update them each time
the project moves on to a new version number or other, not as clearly appearing
differences.

See commits 065f7e4177408536f572253257c1afeded751d79
or ff88bae4d74c24c845371e5e1ebd1f8aff971edf for reference.

Ultimatively, I see no reason to keep them besides the feature of
setting a project specific coding style. This, however, does not convice me
to keep the slightest of those files. Having a clean repository, that focuses on
the source code and the dependencies alone, seems to be the much more valuable
gain. At least for me at this point in time.