Change generation step of code images to ceil square root
Instead of using the floor of the square root of code length, now the ceil is used.
This prevents loss of information on the raw image (befor scaling).
Consider e.g. the predicade "(x>y) & (y>x)" with 13 characters.
rounding the square root down yields 3, so a 3*3 image only hold information
of 9 of 13 characters. Ceiling yields 4, so the 4*4=16 pixel image contains the
predicate fully.

Extra characters are set as space character (ASCII 32).