Make AdjacencyList invariant to surrounding spaces
Adding an identifier "a " made id not associatable as identifier "a".
Now the id's in the AdjacencyList are invariant of surrounding white
space.

Further, specifying domain-boundaries for an identifier resulted in a
null pointer exception if the identifier was not existent before.
Correct it to silently add the id beforehand.