Migrate PredicateCollector from thesis
Changed it to PredicateCollection, as this name better captures what
class instances are. A PredicateCollection collects all predicates from
a given StateSpace, but only collects once. Thus for each new
StateSpace, a new collection is of need. A collector would semantically
be able to collect data of multiple state spaces without having to
create a new instance.