Add weakest precondition for whole invariant
Additionally to only generation the weakest preconditions for each
operation/invariant pair, it is now additionally generated for the
concatenation of all invariants (for each operation still).

In the Master's Thesis the weakest precondition used for the whole
invariant was simply the concatenation of the individual ones of each
invariant. This however was not correct, as due to the concatenation
some additional restriction could be posed, resulting in a stronger
invariant after all. This approach is more suitable to be correct.