Migrate PredicateEvaluator logic to core package
Change SovlerType enum to Backend interface.
This allows for bundling the state space communication logic directly
into the implementing backend. As in the future it might be of interest
to use different settings for the backends as well, a simple enum seems
insufficient, as the possible enumerations might explode exponentially.
Further the addition of new backends comes with simply implementing the
interface, which can be done by any user, instead of expanding the
interface, which needs to be done in the NeuroB source code.

Further migrate logic from PredicateEvaluator into the new Backend
interface. As the implemented backends are now capable to implement
logic of deciding predicates themselves, the PredicateEvaluator logic
can cleanly be moved to them instead of unnecessarily extracting it.