Add Z3BackendIT to verify whether Z3 is callable
If Z3 is not correctly configured on the executing system, no predicates
can be decided by it. This test queries a trivial formula to Z3, and
when it cannot be decided this might indicate a missing Z3 binding.

This is meant to serve as a reference for other integration tests that
might make use of this backend and expect certain outputs. If this test
fails, this means other test results might not be valid.