Fix logic of laziness for training set generation
Instead of not(lazy or exists), training data should be created when
not(lazy) or not(exists) = not(lazy or exists).
This ensures, that if the operation is not lazy, the data will be
created either way. If the operation is lazy, then the generation
depends on whether the file already exists or not.