Remove EventB comments, add parenthesis after '=>'
Weakest Preconditions as well as before/after predicates generated by
ProB contain a /* @example:grd1 */ comment, which actually makes them
unparseable by the B Parser.

Although this is probably undesired behaviour and should be fixed inside
of the ProB core, a work around is still of need

Additionally, wrap parts of the predicates in parenthesis if they follow an
implication ('=>'). As said parts might contain an implication as well,
this could cause errors at parse time. The parser disallows using
implications in a row, as the precedence of operations is not obvious.