Generalise TrainingSample.java to hold BData instead of Features
This is a biggie.  The BData class was introduced as a common generalisation of B
predicates and features. As the predicate data bases only contain
predicates, this allows to use BPredicate directly, instead of misusing
PredicateFeatures to encapsulate a predicate everywhere compatibility
for features was necessary.

Next step was to generalise the formats, so they can state which data
are to be expected (BPredicates, Features, ...) and made switching
between e.g. CsvFormat and JsonDbFormat at training data generation
easier. This inlined the formats more, as instead of having to handle
DbSample objects, they could be substituted by TrainingSample objects
respectively, as both now could hold the same data.

Finally, this substitution was integrated, switching instances of
DbSample usage with TrainingSamples.