Fix FormulaeGenerator null pointer exception
When an invariant was not correctly primed, the primed version is still
tried to be accessed. On this, #toString is called, which results in a
NPE. The null check however is right after.

Instead of calling #toString, the primed invariant can remain a
BPredicate. This would be sorted out by the null check and everything
should work.