Change #getMachineAccess to #spawnMachineAccess
BMachine instances carrying an internal machine access with getter
results in uncertainty about when to call #closeMachineAccess on them.
Root cause was that no code fragment calling the getter actually was
able to safely assume the access was closeable, as it could be accessed
in a subsequent function call.

Changing this behaviour to be more factory like, spawning always a
fresh, new access, establishes a more sophisticated workflow of
he who calls it has to close it.