Increase performance of analyses
Main culprit of slow performance was the multi-label classification.
Adding a sample to the multi-label class {a, b, c} for the labels a, b,
c resulted in generating all subsets of {a,b,c} (7 elements as the empty
set was ignored), using each such set as label and increasing their
counts individually. This had a quadratic run-time and slowed the
analysis down.

Now, adding a sample to the multi-label {a, b, c} does no longer
internally increase the count for all its true subsets too (e.g. the
internal count of {a}, {b}, ... remains the same). However, upon
querying the amount of samples seen belonging to, e.g., {a}, the counts
of each seen subset of classes containing a are summed up, and the
result is returned accordingly.
I.e. #samples a = count({a}) + count({a, b}) + count({a, c}) + ...
As the results of an analysis are usually queried at the very end of
analysis, this reduces the runtime by a large margin.

For the regression analysis, the average is no longer computed on the
fly but rather at query time, much for the same reasons (although the
performance gain hereby is not humanly experienceable).